refactor: move the option_spec to its own file and generalize argument handling
This is a preparation for #16. Doing this refactor I learnt that cluttex
(and thus also cluttealtex) parses arguments which have a default value
quite weird.
If default is set and parameter is specified as --opt=param, the
parameter from the argument is taken.
If default is set and parameter is specified as --opt param, the default
paramter is taken and the cli parameter is ignored (probably produces an
error later during parsing).

The main part of the refactor was
1. move the option_spec to its own file
2. move handling the options inside the option_spec table
3. generate LUT to avoid having to iterate over the complete table

Now when adding a new option all that needs to be done is in the
option_spec.lua file (adding the entry in the table and appending to the
usage output).