Rewriting the docs showed the names of the filers for the watcher are misleading
The keywords `only`/`not` give the options a very final sounding. Also
`only` is not really true depending on the setting (depends on what the
previous filters did).

Thus, these keywords in the options are renamed to
`inc`(lude)/`exc`(lude). In order to avoid confusion with the `inc-exc`
option (which otherwise would really close to `inc-ext`), this option
is also renamed to `filter`.